update the logic related to characters.
place in a separate script the logic of movement, changing the direction of view and animation  that is responsible for the actions of the enemy's body (EnemyBody). The same logic is implemented for the main character.
For more flexibility, CharacterBody is a separate MonoBehavior component.

The scripts responsible for the characters were refactored.

All prefabs have been updated to reflect the changes.

Translated with www.DeepL.com/Translator (free version)