implement the Attack Impact logic, create new enemy type, update files structure.
implement the attack impact for the character sword and for default enemy weapon:
impact includes a knockback and a stun effect + the visual target flash effect,
there are certain chance that impact will be performed and the reload time,
add flash shader and update the JitterFreeLit to implement the flash effect,
add the EntityPhysics component for each entity to perform physics actions on it,
implement a EntityPhysicsActions script to create custom sets of physics impacts + reuse them.

update the entity data asset logic :
move the prefab to the spawn data and refactore all related scripts,
use the inspector to assign main character data for tests.

add a new enemy type - BioMonster and create all prefabs/data assets for him.
update the structure of scripts in the root folder, move interfaces and SO to the common folders.
add some test assets (chest) for testing.