implement damage indicators logic, add new services, update DI installers.
add EnemyManagementService to implement some additional actions related to the enemies.
add GameplayIndicatorsService to create and manage text indicators using the object pool.
create IndicatorsPool and update EnemyPool.
create Text Indicator prefab with the configured TMPro component and with the custom material (+a texture).
create the animation and the animation event for the text indicator to hide it slowly.
create the GameplayServiceManager and the StageManager to configure the basic services for certain scenes.
update the damage logic + create new structures for attacks/effects/damage data.
update the entities logic with the new damage info and all related scripts (weapons, etc.)
update the enemy's return logic (use EnemySpawner to return the enemy instead of the direct pool link).
update the EnemyConfigurator : add a method to deconfigure the enemy while returning to the spawner.
rename some scripts and SO (rename some services named as managers to be "Service").