implement the character level system, add drop items, implement the wind and the emission effects, fix some bugs.
implement the character level system :
implement the calculation of the character level bounds,
implement methods to add the xp and update levels,
implement the CharacterLevelData SO to configure level main settings.
create the UI elements for character level and configure smooth animations (for the xp bar),
remove the character from the Gameplay_Essentials_Scene and use the CharactersInstaller to load and create him,
update all related scripts.

add drop items :
create some base classes/interfaces and data assets for drop items,
implement the BasicItemXp, the ItemXpData and the ItemXpParameters,
implement the ItemDropData to configure items and their drop chances.
configure the EnemyData to contain the drop items list.
configure the EnemyController to call the spawn of random item when enemy is dead.
create some data assets and prefabs for drop items (for tests),

implement wind materials and emission materials :
create the wind shadergraph and some configured materials based on it,
create GrassVelocityController and GrassExternalVelocityTrigger to apply the character's velocity to the grass.
create the emission shadergraph (for future visual effects).

fix small bugs :
fix the HealthBar side switch (maybe not fixed completely),
fix the character stun (movement stop) and the attack (movement stop),
for details check the Trello card -> small-fixes-#1.