implement the pause menu, implement HealCapsule and DarkPillar, add the emission shader, etc.
implement the pause menu with all related logic :
implement the pause logic to stop the whole gameplay,
implement the game pause when the ESC key or the UI button is pressed,
implement the pause menu UI with the video background and with the fade-in effect (add BezierUtils for that),
implement the basic logic for some buttons (continue/exit).

implement the HealCapsule with 3 different types (red, blue, golden) :
implement the red capsule with the instant health regen,
implement the blue capsule with the temporary slow health regen,
implement the golden capsule with the permanent health regen,
add some health capsule assets to the enemy drop items.

implement the DarkPillar with all animations, use the temp script to test all animations,
configure all transitions between animations in Animator,
create the emission temp shader (using SG) and apply it to the pillar.

sort some assets/scripts in the root folder.
run the code cleanup for all scripts (to build and test the game correctly).