implement the main menu, import the new font, fix some issues with Addressable Assets.
import the Dofded font and use it for some UI elements.
implement the Initialization_Scene and add it to the build.
configure the AddressableSceneLoader and GameSceneManager.

implement the main menu with the start/exit buttons, the settings menu, the credits menu.
implement empty panels for settings and credits, configure the buttons logic to open them.
import Background_Video for main menu and configure it.
update all related scripts.