implement a basic upgrades system with flexible upgrades configuration.
create UI cards for upgrades and configure their animations.
create a basic logic to display cards with upgrades and apply them to the target.
configure the SO hierarchy to create different upgrades and reuse existing SO for them.
create some data assets for upgrades for tests (CyberHeart)
create some single upgrades : CharacterSpeedUpgradeSO/CharacterHealthUpgradeSO.
create a main UpgradeConfigurationSO to configure the upgrade data and UI visual.

add a script to the restart button in the pause menu.
fix the scene loading issue : do not use "activateOnLoad = false"
fix the loading screen percentage update issue : calculate the correct update duration.
fit the character level UI issue (use event args instead of the direct link).
fix the performance issue :
initialize GameSavePaths in the GameInitializer,
remove the logic which saves stage stats to the json when game is finishing (need an async method for that).