update the upgrade system logic, create some basic upgrade configurations, implement AddressableAssetsHolder.
create AddressableAssetsHolder to cache and reuse loaded Addressable assets,
refactore EntityFeaturesHolderBase and concrete holders to reuse cached assets instead of loading duplicates,
add an ID for each EntityWeaponData and FeatureData, use them in holders.

implement the logic for unlocking weapons and provide a special UpgradeConfigurationSO to the UpgradesController for unlocked weapons,
implement the logic for applying upgrades to the specific weapons (using these special Configurations and weapon's ID),
implement UpgradeProvider and use it instead of concrete SO.
add the upgrade info to SingleUpgradeBaseSO and use it in UpgradeProvider to give the specific info about current upgrade level.

create UpgradesSetData for default set of upgrades and use DI to inject it on the level start,
create several basic configurations for unlocking weapons and upgrading them (CyberSword and ElectroBlade),
create several basic configurations for upgrading the character and testing different upgrade types (CyberWalk and CyberHeart).

import all available icons (128px) for upgrades and create a sprite atlas for them.